ref(issues): Remove old UI from remaining event interface files
